requirements.txt
flask==3.0.2
python-dotenv==1.0.1
llama-index==0.12.26
openai==1.12.0
werkzeug==3.0.1
setuptools==68.2.2 